+++ src/modules/identity/users/types/public-user.type.ts
@@ -1,10 +1,11 @@
-import { UserRole, UserStatus } from '@prisma/client';
+import { AccountType, UserRole, UserStatus } from '@prisma/client';
 
 export type PublicUser = {
   id: string;
   email: string;
   username: string;
   role: UserRole;
+  accountType: AccountType;
   status: UserStatus;
   reputation: number;
   createdAt: Date;